+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+### Changed
+- duration of proposal phase is self-adjustable
+- staking token percentage is keep for individual election
+
 ## [0.5.0] - 2019-10-29
 ### Added
 - flattener script